+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
 
 - ci(): add source map support to node sandbox [#9686](https://github.com/fabricjs/fabric.js/pull/9686)
+- feat(Circle): Add counterclockwise parameter to Circle class [#9670](https://github.com/fabricjs/fabric.js/pull/9670)
 
 ## [6.0.0-beta19]