+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
 
 - fix(): first touch gesture [#9684](https://github.com/fabricjs/fabric.js/pull/9684)
+- feat(Circle): Add counterclockwise parameter to Circle class [#9670](https://github.com/fabricjs/fabric.js/pull/9670)
 
 ## [6.0.0-beta19]